% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bigdat}
\alias{bigdat}
\title{A simulated multistate dataset with lots of observations and
covariates}
\format{
See \code{infsim/bigdata.R} in the source for simulation settings.
}
\usage{
bigdat
}
\description{
A simulated multistate dataset with lots of observations and
covariates
}
\keyword{datasets}
