% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{edf}
\alias{edf}
\title{Misclassification error probabilities from an msmbayes model}
\usage{
edf(draws)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}
}
\value{
A data frame with one row per modelled misclassification probability.
\code{from} indicates the true state, and \code{to} indicates the observed state.
Error probabilities fixed by the user are not included.
}
\description{
Misclassification error probabilities from an msmbayes model
}
\seealso{
\code{\link{qdf}} for more information about the format.
}
