% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{ematrix}
\alias{ematrix}
\title{Matrix of misclassification error probabilities from an msmbayes model}
\usage{
ematrix(draws)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}
}
\value{
An array or matrix of \code{rvar} objects containing the
misclassification error matrix for each new prediction data point
}
\description{
Matrix of misclassification error probabilities from an msmbayes model
}
