% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_models}
\alias{example_models}
\alias{infsim_model}
\alias{infsim_modelc}
\alias{cav_misc}
\title{Example fitted model objects used for testing msmbayes}
\format{
An object of class \code{msmbayes}, obtained by fitting a
Markov model to the dataset \code{\link{infsim}}.  See
\code{data-raw/infsim.R} in the source for the model
specification code

\code{infsim_modelc} includes covariates on the transition
intensities.

\code{cav_misc} is the misclassification model fitted to the CAV data from msm in the "Advanced multi-state models in msmbayes" vignette.
intensities.
}
\source{
Simulated
}
\usage{
infsim_model

infsim_modelc

cav_misc
}
\description{
Example fitted model objects used for testing msmbayes
}
\keyword{datasets}
