% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{gamma_shape_in_bounds}
\alias{gamma_shape_in_bounds}
\title{Test whether a shape parameter of the Gamma is in the bounds
required for a valid phase-type approximation}
\usage{
gamma_shape_in_bounds(shape, nphase)
}
\arguments{
\item{shape}{Shape parameter or vector)}

\item{nphases}{Number of phases}
}
\value{
Vector or logicals, whether each shape parameter is in the bounds
require for a phase-type approximation with that number of phases.
}
\description{
Test whether a shape parameter of the Gamma is in the bounds
required for a valid phase-type approximation
}
\details{
Also verifies that the parameter satisfies Case 1 of Theorem 1
in Bobbio et al.
}
