% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{hr}
\alias{hr}
\title{Hazard ratios for covariates on transition intensities}
\usage{
hr(draws)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}
}
\value{
A data frame containing samples from the posterior distribution.
See \code{\link{qdf}} for notes on this format and how to summarise.
}
\description{
Hazard ratios for covariates on transition intensities
}
\seealso{
\code{\link{loghr}}
}
