% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{illdeath_data}
\alias{illdeath_data}
\title{A simulated dataset from an illness-death model}
\format{
See \code{data-raw/illdeath.R} in the source for simulation settings.
}
\usage{
illdeath_data
}
\description{
A simulated dataset from an illness-death model
}
\keyword{datasets}
