% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{infsim}
\alias{infsim}
\alias{infsim2}
\title{Simulated infection testing data}
\format{
\code{infsim} has 3600 rows, with 36 state observations for each of 100 people.  A smaller dataset \code{infsim2} has only 360 rows, from 20 people, and
is simulated using a sojourn time of 60 days in the test-negative state
and 10 days in test-positive.

Columns are:
\itemize{
\item \code{subject} Subject identifier
\item \code{days} Observation time (in days)
\item \code{months} Observation time (in moths)
\item \code{state} State simulated from a Markov model with no covariates
\item \code{sex}: "male" or "female".
\item \code{age10}: Age, in units of 10 years since age 50
\item \code{statec}: State simulated from a Markov model with covariates
\item \code{statep}: State simulated from a phase-type model (unused in any examples. See \code{data-raw/infsim.R} in the source for simulation settings)
\item \code{statepc}: State simuated from a phase-type model with covariates (unused in any examples)
}

An object of class \code{data.frame} with 360 rows and 14 columns.
}
\source{
Simulated
}
\usage{
infsim

infsim2
}
\description{
Simulated infection testing data
}
\details{
The transition intensities used for the simulation are defined
using a mean sojourn time of 180 days in the "test negative" state
and 10 days in the "test positive" state.

For the model with covariates, the log hazard ratios are
2 for male on the 1-2 transition, 1 for \code{age10} on the 1-2 transition,
and -1 for \code{age10} on the 2-1 transition.  Baseline intensities are
for female, age 50 (i.e. \verb{age10=} 0).

In the state data, state 1 is negative and 2 is positive.
}
\keyword{datasets}
