% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{infsim_model}
\alias{infsim_model}
\alias{infsim_modelc}
\title{Example fitted model objects used for testing msmbayes}
\format{
An object of class \code{msmbayes}, obtained by fitting a
Markov model to the dataset \code{\link{infsim}}.  See
\code{data-raw/infsim.R} in the source for the model
specification code

\code{infsim_modelc} includes covariates on the transition
intensities.
}
\source{
Simulated
}
\usage{
infsim_model

infsim_modelc
}
\description{
Example fitted model objects used for testing msmbayes
}
\keyword{datasets}
