% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{loabs_pars}
\alias{loabs_pars}
\title{Summarise posteriors for log odds of transitions from phase-type
states}
\usage{
loabs_pars(draws)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}
}
\description{
Log odds of transition to a competing destination state, relative
to baseline destination state.  Only applicable to phase-type
approximation models, specified with \code{pastates}.
}
