% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{loghr}
\alias{loghr}
\alias{hr}
\title{(Log) hazard ratios for covariates on transition intensities}
\usage{
loghr(draws)

hr(draws)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}
}
\value{
A data frame containing samples from the posterior distribution.
See \code{\link{qdf}} for notes on this format and how to summarise.
\code{\link{hr}} returns these on the natural scale, \code{\link{loghr}}
returns them on the log scale.
}
\description{
In semi-Markov models specified with \code{pastates}, these only include
covariate effects on transitions out of "Markov" states, that is, those not included
in \code{pastates}.  Effects on scale parameters of sojourn distributions
in "semi-Markov" states are extracted with \code{\link{logtaf}}.
}
