% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{loglik}
\alias{loglik}
\alias{logLik.msmbayes}
\title{Log likelihood from an msmbayes model}
\usage{
loglik(draws)

\method{logLik}{msmbayes}(object, ...)
}
\value{
A data frame with columns indicating

\code{posterior}: the posterior distribution of the log likelihood (as an \code{rvar} object).

\code{mode} the posterior mode of the log likelihood (only if optimisation was used to fit the model, rather than MCMC).  If \code{msmbayes} was called \code{priors="mle"}, this is the maximised log likelihood.
}
\description{
Log likelihood from an msmbayes model
}
