% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{logtaf}
\alias{logtaf}
\alias{taf}
\title{(Log) time acceleration factors in semi-Markov models}
\usage{
logtaf(draws)

taf(draws)
}
\value{
A data frame containing samples from the posterior distribution.
See \code{\link{qdf}} for notes on this format and how to summarise.

\code{\link{taf}} returns these on the natural scale, \code{\link{logtaf}}
returns them on the log scale.
}
\description{
Extracts the covariate effects on scale parameters in phase-type approximation
sojourn distributions in semi-Markov models.   Note these are not hazard
ratios for transitions on the observable state space.  They are time acceleration
factors, such that an increase in one covariate unit makes time pass at \code{taf}
times the speed, such that the sojourn time is expected to reduce by \eqn{exp(-taf)}.
The log TAFs are labelled \eqn{\beta} in the paper and vignettes.
}
\details{
See \code{loghr} for effects of covariates on transitions out of Markov states,
which are log hazard ratios.
}
