% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{mean_sojourn}
\alias{mean_sojourn}
\title{Mean sojourn times from an msmbayes model}
\usage{
mean_sojourn(draws, new_data = NULL, by_phase = TRUE)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{new_data}{Data frame with covariate values to predict for}

\item{by_phase}{For states with phase type distributions:

If \code{TRUE} then one mean sojourn time per phase is returned.

If \code{FALSE}, then one overall mean sojourn time for the state
is returned.  This is not generally equal to the sum of the
phase-specific mean sojourn times, because an individual may
transition out of the state before progressing to the next phase.}
}
\value{
A data frame containing samples from the posterior distribution.
See \code{\link{qdf}} for notes on this format and how to summarise.
}
\description{
Mean sojourn times from an msmbayes model
}
