% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseapprox_msm.R
\name{msm_phaseapprox}
\alias{msm_phaseapprox}
\title{Maximum likelihood estimation of multi-state models for intermittently-observed data where one state has a non-exponential distribution approximated by a phase-type model}
\usage{
msm_phaseapprox(
  formula,
  subject,
  data,
  qmatrix,
  phased_state,
  family = "weibull",
  par,
  spline = "linear",
  fit_method = "optim",
  control = NULL
)
}
\arguments{
\item{formula}{model formula as in msm}

\item{subject}{unquoted name as in msm}

\item{data}{data frame}

\item{qmatrix}{0/1 transition structure as in msm}

\item{phased_state}{integer state given phase type distribution. Only one phased state permitted}

\item{family}{parametric family approximated by the phase-type distribution: \code{"weibull"} or \code{"gamma"}}

\item{par}{initial value vector. shape on natural scale, rest on log scale}

\item{spline}{(advanced) \code{"linear"} or \code{"hermite"}}

\item{fit_method}{\code{"optim"} or \code{"SGA"}}

\item{control}{passed to fit method}
}
\value{
list of optimisation results TODO refine
}
\description{
Maximum likelihood estimation of multi-state models for intermittently-observed data where one state has a non-exponential distribution approximated by a phase-type model
}
