% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmbayes-package.R
\docType{package}
\name{msmbayes-package}
\alias{msmbayes-package}
\title{The 'msmbayes' package.}
\description{
For an introduction to and overview of the \code{msmbayes} package, and full documentation, see

\url{GITHUB TODO}.

For details of the methods, see TODO
}
\references{
Kalbfleisch and Lawless, Statn.
}
\author{
\strong{Maintainer}: Christopher Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk} (\href{https://orcid.org/0000-0002-6656-8913}{ORCID})

}
