% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmhist.R
\name{msmhist_bardata}
\alias{msmhist_bardata}
\title{Estimate state occupation probabilities to be illustrated by a bar
plot in \code{msmhist}}
\usage{
msmhist_bardata(
  data,
  state,
  time,
  subject,
  nbins,
  absorbing = NULL,
  censtimes = NULL
)
}
\arguments{
\item{data}{Data frame giving the observed data.}

\item{state}{Character string naming the observed state variable in
the data.  This variable must either be an integer in 1,2,...,K,
where K is the number of states, or a factor with these integers
as level labels.  If omitted, this is assumed to be \code{"state"}.}

\item{time}{Character string naming the observation time variable in the data.
If omitted, this is assumed to be \code{"time"}.}

\item{subject}{Character string naming the individual ID variable in the data.
If omitted, this is assumed to be \code{"subject"}.}

\item{nbins}{Number of time intervals to bin the state observations
into.  The underlying distribution of states illustrated by the
plot will be assumed constant within each interval.}

\item{absorbing}{Indices of any absorbing states.  Individuals are
assumed to stay in their absorbing state, and contribute one
observation to each bin after their absorption time.  By default,
no states are assumed to be absorbing.}

\item{censtimes}{Vector of maximum intended follow-up times for the
people in the data who entered absorbing states.  This supposes
that had the person not entered the absorbing state, they would
not have been observed after this time.}
}
\value{
Data frame with one row per bin and state, and columns:
\itemize{
\item \code{binid}: Integer ID for bin
\item \code{binlabel}: Character label for bin, with time interval
\item \code{state}: State
\item \code{binstart}, \code{binend}: Start and end time of the bin (numeric)
\item \code{props}: estimates of state $s$ occupancy proportions $p(s)$ for each bin
\item \code{cumpstart}, \code{cumpend}: Cumulative sum of \code{props} over the set of
states, where \code{cumpstart} starts at 0, and \code{cumpend} ends at
\enumerate{
\item Intended for creating stacked bar plots with \code{geom_rect} or
similar.
}
}
}
\description{
Estimate state occupation probabilities to be illustrated by a bar
plot in \code{msmhist}
}
\seealso{
\code{\link{msmhist}}
}
