% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{n3_moment_bounds}
\alias{n3_moment_bounds}
\title{Bounds on normalised moments for phase-type approximations}
\usage{
n3_moment_bounds(n2, n)
}
\arguments{
\item{n2}{Second normalised moment}

\item{n}{Number of phases}
}
\value{
List with components \code{lower}, \code{upper} defining the
bounds on the third normalised moment (\code{n3}) required for
\code{n2} and \code{n3} to be the moments of a phase type distribution
with \code{n} phases.
}
\description{
From Bobbio et al. (Theorem 3.1)
}
