% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphase.r
\name{nphase_Q}
\alias{nphase_Q}
\title{Given a phase-type sojourn distribution, return the corresponding
Markov intensity matrix where the last state is the absorbing
state, and the the time to absorption is the sojourn distribution.}
\usage{
nphase_Q(prate, arate)
}
\arguments{
\item{prate}{Progression rates.  Either a vector of length
\code{nphase-1}, or a matrix with \code{npar} rows and \code{nphase-1} columns.}

\item{arate}{Absorption rates.  Either a vector of length \code{nphase},
or a matrix with \code{npar} rows and \code{nphase} columns.}
}
\description{
Given a phase-type sojourn distribution, return the corresponding
Markov intensity matrix where the last state is the absorbing
state, and the the time to absorption is the sojourn distribution.
}
