% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{padest_pars}
\alias{padest_pars}
\title{Probabilities of competing exit destination states in phase type
models}
\usage{
padest_pars(draws)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}
}
\value{
A dataframe with one row per probability, and colums for
(from) state, to-state, posterior distribution (as an \code{rvar}
object) and posterior mode (if the model was fitted by mode
optimisation).

padest stands for probability of absorption (from phase system) in
destination state ... TODO clearer name?  pcomprisk? pnext? pnextstate? we also
have loa, logoddsabs
}
\description{
These are the inverse multinomial logit transforms of the
log odds of transition, relative to the first potential exit state.
}
