% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseapprox.R
\name{phase5approx}
\alias{phase5approx}
\title{Training data to construct phase-type approximations of standard survival distributions}
\source{
Code in kl_pointwise.R to provide...
}
\usage{
phase5approx(family)
}
\arguments{
\item{family}{parametric family approximated by the phase-type distribution: \code{"weibull"} or \code{"gamma"}}
}
\value{
\code{traindat}: data frame of training data giving the optimal
phase-type parameters for each of a grid of shape parameters
TODO rename a to shape

Do we need anything else?  gradient?

smods is redundant surely

TODO document further, work in progress.
}
\description{
Only 5-phase approximations are included.
}
