% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{pmatrixdf}
\alias{pmatrixdf}
\title{Transition probabilities from an msmbayes model, presented
as a tidy data frame}
\usage{
pmatrixdf(draws, t = 1, new_data = NULL)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{t}{prediction time or vector of prediction times}

\item{new_data}{Data frame with covariate values to predict for}
}
\value{
A data frame containing samples from the posterior distribution.
See \code{\link{qdf}} for notes on this format and how to summarise.
}
\description{
Transition probabilities from an msmbayes model, presented
as a tidy data frame
}
