% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{pnext_phaseapprox}
\alias{pnext_phaseapprox}
\title{Probabilities of competing exit destination states in phase type
models}
\usage{
pnext_phaseapprox(draws)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}
}
\value{
A dataframe with one row per probability, and colums for
(from) state, to-state, posterior distribution (as an \code{rvar}
object) and posterior mode (if the model was fitted by mode
optimisation).

In models with covariates on the transition odds, this currently only presents
these parameters for covariate values of zero.

TODO rename loa, logoddsabs to include "next" ?
}
\description{
These are the inverse multinomial logit transforms of the
log odds of transition, relative to the first potential exit state.
}
