% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{qmatrix}
\alias{qmatrix}
\title{Transition intensity matrix from an msmbayes model}
\usage{
qmatrix(draws, new_data = NULL, X = NULL, drop = TRUE, type = "posterior")
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{new_data}{Data frame with covariate values to predict for}

\item{X}{Lower-level alternative to specifying \code{new_data}, for
developer use only.  \code{X} is a numeric matrix formed from
column-binding the covariate design matrices for each transition
in turn.}

\item{drop}{Only used if there are no covariates supplied in
\code{new_data}. Then if \code{drop=TRUE} this returns a
\code{nstates} x \code{nstates} matrix, or if \code{drop=FALSE}
this returns a 3D array with first dimension \code{ncovs=1}.}

\item{type}{\code{"posterior"} to return \code{rvar} objects containing posterior samples.

\code{"mode"} to return posterior modes (only applicable if model was fitted
with posterior mode optimisation).}
}
\value{
An array or matrix of \code{rvar} objects or numbers, representing the
transition intensity matrix for each new prediction data point
}
\description{
Transition intensity matrix from an msmbayes model
}
\examples{
qmatrix(infsim_model)
summary(qmatrix(infsim_model))
summary(qmatrix(infsim_model), median, ~quantile(.x, c(0.025, 0.975)))

}
\seealso{
\code{\link{qdf}} returns the same information in a
tidy data frame format
}
