% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{rrnextdoc}
\alias{rrnextdoc}
\alias{logrrnext}
\alias{rrnext}
\title{Effects of covariates on competing exit transitions in phase type
models}
\usage{
logrrnext(draws)

rrnext(draws)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}
}
\value{
A data frame containing samples from the posterior distribution.
See \code{\link{qdf}} for notes on this format and how to summarise.
}
\description{
Effects of covariates on competing exit transitions in phase type
models
}
\details{
Only applicable to phase-type approximation models,
specified with \code{pastates}.

\code{logrrnext} gives the Linear effect of covariates on log relative
risk of transition to a competing destination state, relative to
baseline destination state.

\code{rrnext} gives the exponential of the linear effect,
interpretable as a hazard ratio.  See the semi-Markov models
vignette and paper for the mathematical details.
}
