% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{shape_ubound}
\alias{shape_ubound}
\title{Upper bound for shape parameter in moment-based phase-type approximations}
\usage{
shape_ubound(nphase, family)
}
\arguments{
\item{nphase}{Number of approximating phases}

\item{family}{\code{"weibull"} or \code{"gamma"}}
}
\value{
Upper bound for shape parameter
}
\description{
Upper bound for shape parameter in moment-based phase-type approximations
}
