% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_2state_smm}
\alias{sim_2state_smm}
\title{Simulate intermittently-observed data from a semi-Markov
multi-state model with two states and reversible transitions.}
\usage{
sim_2state_smm(
  nindiv,
  obstimes,
  rfn1 = rexp,
  pars1 = list(rate = 1),
  rfn2 = exp,
  pars2 = list(rate = 1)
)
}
\arguments{
\item{nindiv}{Number of individuals.}

\item{obstimes}{Observation times, common between individuals.}

\item{rfn1}{Function to simulate from the sojourn distribution in
state 2.  By default this is the exponential.}

\item{pars1}{Named list of arguments and their values to be passed
to \code{rfn1}, specifying parameter values for the sojourn
distribution.}

\item{rfn2}{Function to simulate from the sojourn distribution in
state 2.}

\item{pars2}{Named list of arguments and their values to be passed
to \code{rfn2}, specifying parameter values for the sojourn
distribution.}
}
\description{
Either or both states can have any sojourn distribution that we
know how to simulate from.
}
