% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{soj_prob}
\alias{soj_prob}
\title{Sojourn probability in a state of a msmbayes model}
\usage{
soj_prob(draws, t, state, new_data = NULL, method = "analytic")
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{t}{Time since state entry.  A single time or a vector can be supplied.}

\item{state}{State of interest (A single integer)}

\item{new_data}{Data frame with covariate values to predict for}

\item{method}{Only applicable to phase-type models. Method for computing
the matrix exponential involved in the phase-type sojourn distribution.
See \code{\link{pnphase}}.}
}
\value{
A data frame with column \code{posterior} giving the posterior
distribution for the probability of remaining in \code{state} by time
\code{t} since state entry, as an \code{rvar} object. Other columns give
the time and any covariate values.

See \code{\link{qdf}} for notes on the \code{rvar} format.
}
\description{
Sojourn probability in a state of a msmbayes model
}
