% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{soj_quantile}
\alias{soj_quantile}
\title{Quantiles of the sojourn distribution in a state of a msmbayes model}
\usage{
soj_quantile(draws, p, state, new_data = NULL, method = "analytic")
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{p}{Vector of probabilities at which to evaluate quantiles of the sojourn distribution}

\item{state}{State of interest (A single integer)}

\item{new_data}{Data frame with covariate values to predict for}

\item{method}{Only applicable to phase-type models. Method for computing
the matrix exponential involved in the phase-type sojourn distribution.
See \code{\link{pnphase}}.}
}
\value{
A data frame with column \code{posterior} giving the posterior
distribution (as an \code{rvar} object) for the time in \code{state} for
which the probability of remaining in this state by this time is
\code{p}.
}
\description{
Quantiles of the sojourn distribution in a state of a msmbayes model
}
\details{
For the inverse of this function, see \code{\link{soj_prob}}.
}
