% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{standardise_to}
\alias{standardise_to}
\alias{standardize_to}
\title{Constructor for a standardising population used for model
outputs}
\usage{
standardise_to(new_data)

standardize_to(new_data)
}
\arguments{
\item{new_data}{Data frame with covariate values to predict for}
}
\description{
Standardised outputs are outputs from models with covariates, that
are defined by marginalising (averaging) over covariate values in a
given population, rather than being conditional on a given
covariate value.
}
\details{
Standardised outputs are produced from a Monte Carlo sample from
the joint distribution of parameters \eqn{\theta} and covariate
values \eqn{X}, \eqn{p(X,\theta) = p(\theta|X)p(X)}, where
\eqn{p(X)} is defined by the empirical distribution of covariates
in the standard population.   This joint sample is obtained by
concatenating samples of covariate-specific outputs.

Hence applying an output function \eqn{g()} (such as the transition
probability) to this sample produces a sample from the posterior of
\eqn{\int g(\theta|X) dX}: the average transition probability (say)
for a heterogeneous population.
}
\examples{

nd <- data.frame(sex=c("female","male"))

## gender-specific outputs
qdf(infsim_modelc, new_data = nd) 

## averaged over men and women in the same proportions as are in `nd`
## in this case, `nd` has two rows, so we take a 50/50 average
qdf(infsim_modelc, new_data = standardise_to(nd))

}
