% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{totlos}
\alias{totlos}
\title{Total length of stay in each state over an interval}
\usage{
totlos(
  draws,
  t,
  new_data = NULL,
  fromt = 0,
  pstart = NULL,
  discount = 0,
  states = "obs"
)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{t}{End point of the time interval over which to measure
length of stay in each state}

\item{new_data}{Data frame with covariate values to predict for}

\item{fromt}{Starting point of the time interval, by default 0}

\item{pstart}{Vector giving distribution of states at time 0}

\item{discount}{Discount rate in continuous time}

\item{states}{If \code{states="obs"} then this describes mean sojourn times in
the observable states.  For phase-type models this is not
generally equal to the sum of the phase-specific mean sojourn
times, because an individual may transition out of the state
before progressing to the next phase.

If \code{states="phase"} then for phase-type models, this describes mean sojourn times
in the latent state space.}
}
\value{
Data frame with one row for each state and covariate value,
giving the expected amount of time spent in that state over the
forecast interval.
}
\description{
See \code{\link[msm:totlos.msm]{msm::totlos.msm()}} for the theory behind the method used to
calculate this.  The analytic formula is used, not numerical integration.
}
