% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{ematrix}
\alias{ematrix}
\title{Matrix of misclassification error probabilities from an msmbayes model}
\usage{
ematrix(draws, type = "posterior")
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{type}{\code{"posterior"} to return \code{rvar} objects containing posterior samples.

\code{"mode"} to return posterior modes (only applicable if model was fitted
with posterior mode optimisation).}
}
\value{
An array or matrix of \code{rvar} objects containing the
misclassification error matrix for each new prediction data point
}
\description{
Matrix of misclassification error probabilities from an msmbayes model
}
