% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_models}
\alias{example_models}
\alias{infsim_model}
\alias{infsim_modelc}
\alias{infsim_modelp}
\alias{infsim_modelpc}
\title{Example fitted model objects used for testing msmbayes}
\format{
An object of class \code{msmbayes}, obtained by fitting a
Markov model to the dataset \code{\link{infsim}}.  See
\code{data-raw/infsim.R} in the source for the model
specification code

\code{infsim_modelc} includes covariates on the transition
intensities.

\code{infsim_modelp} is a phase-type model.

\code{infsim_modelpc} is a phase-type model with covariates.
}
\source{
Simulated
}
\usage{
infsim_model

infsim_modelc

infsim_modelp

infsim_modelpc
}
\description{
Example fitted model objects used for testing msmbayes
}
\keyword{datasets}
