% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{loglik}
\alias{loglik}
\alias{logLik.msmbayes}
\title{Log likelihood from an msmbayes model}
\usage{
loglik(draws)

\method{logLik}{msmbayes}(object, ...)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{object}{Object returned by \code{\link{msmbayes}}.}

\item{...}{Further arguments passed to both \code{\link{qdf}},
\code{\link{hr}}, \code{\link{loghr}} and \code{\link{edf}}.}
}
\value{
For \code{loglik}, a data frame with rows for the log likelihood, log prior density and log posterior density,
and columns for the posterior (as an \code{rvar} object) and the mode (only if optimisation was used to fit the model, rather than MCMC).

If \code{msmbayes} was called with \code{priors="mle"}, the maximised log posterior and log likelihood should be the same.

For \code{logLik} (note the different capitalisation), just the likelihood (mode if available, or posterior if not) is returned.  This is a
method for the generic \code{logLik} function in the \code{stats} package.
}
\description{
Log likelihood from an msmbayes model
}
