% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{logoddsnext}
\alias{logoddsnext}
\title{Summarise posteriors for log odds of transitions from phase-type
states}
\usage{
logoddsnext(draws, new_data = NULL, keep_covid = FALSE)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{new_data}{Data frame with covariate values to predict for}

\item{keep_covid}{(logical) Keep the integer column \code{covid} identifying unique
covariate combinations.}
}
\description{
Log odds of transition to a competing destination state, relative
to baseline destination state.  Only applicable to phase-type
approximation models, specified with \code{pastates}.
}
\seealso{
pnext
}
