% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{mean_sojourn}
\alias{mean_sojourn}
\title{Mean sojourn times from an msmbayes model}
\usage{
mean_sojourn(draws, new_data = NULL, states = "obs")
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{new_data}{Data frame with covariate values to predict for}

\item{states}{If \code{states="obs"} then this describes mean sojourn times in
the observable states.  For phase-type models this is not
generally equal to the sum of the phase-specific mean sojourn
times, because an individual may transition out of the state
before progressing to the next phase.

If \code{states="phase"} then for phase-type models, this describes mean sojourn times
in the latent state space.}
}
\value{
A data frame containing samples from the posterior distribution.
See \code{\link{qdf}} for notes on this format and how to summarise.
}
\description{
Mean sojourn times from an msmbayes model
}
