% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmbayes-package.R
\docType{package}
\name{msmbayes-package}
\alias{msmbayes-package}
\title{The 'msmbayes' package for Bayesian multi-state modelling of intermittently-observed data.}
\description{
For an introduction to and overview of the \code{msmbayes} package, and full documentation, see

\url{http://chjackson.github.io/msmbayes}.

For more resources on multi-state modelling, see the \href{http://chjackson.github.io/msm}{\code{msm} package} and its documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chjackson/msmbayes}
  \item Report bugs at \url{https://github.com/chjackson/msmbayes/issues}
}

}
\author{
\strong{Maintainer}: Christopher Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk} (\href{https://orcid.org/0000-0002-6656-8913}{ORCID})

}
