% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmbayes.R
\name{msmbayes}
\alias{msmbayes}
\title{Bayesian multi-state models for intermittently-observed data}
\usage{
msmbayes(
  data,
  state,
  time,
  subject,
  Q,
  E = NULL,
  covariates = NULL,
  nphase = NULL,
  priors = NULL,
  fit_method = "sample",
  keep_data = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame giving the observed data.}

\item{state}{Character string naming the observed state variable in
the data.  This variable must either be an integer in 1,2,...,K,
where K is the number of states, or a factor with these integers
as level labels.}

\item{time}{Character string naming the observation time variable in the data}

\item{subject}{Character string naming the individual ID variable in the data}

\item{Q}{Matrix indicating the transition structure.  A zero entry
indicates that instantaneous transitions from (row) to (column)
are disallowed.  An entry of 1 (or any other positive value)
indicates that the instantaneous transition is allowed.  The
diagonal of \code{Q} is ignored.

There is no need to "guess" initial values and put them here, as
is sometimes done in \code{msm}.  Initial values for fitting are
determined by Stan from the prior distributions, and the specific
values supplied for positive entries of \code{Q} are disregarded.}

\item{E}{If \code{NULL} a non-hidden Markov model is fitted.  If
non-\code{NULL} this should be a matrix indicating the structure
of allowed misclassifications, where rows are the true states,
and columns are the observed states.  A zero \eqn{(r,s)} entry
indicates that true state \eqn{r} cannot be observed as observed
state \eqn{s}.  A non-zero \eqn{(r,s)} entry indicates an
initial value for a permitted misclassification probability.  The
diagonal of \code{E} is ignored.}

\item{covariates}{Specification of covariates on transition intensities.
This should be a list of formulae.  Each formula should have a
left-hand side that looks like \code{Q(r,s)}, and a right hand side
defining the regression model for the log of the transition intensity
from state \eqn{r} to state \eqn{s}.

For example,

\code{covariates = list(Q(1,2) ~ age + sex,
                        Q(2,1) ~ age)}

specifies that the log of the 1-2 transition intensity is an additive
linear function of age and sex, and the log 2-1 transition intensity
is a linear function of age.  You do not have to list all of the
intensities here if some of them are not influenced by covariates.}

\item{nphase}{For phase-type models, this is a vector with one
element per state, giving the number of phases per state.  This
element is 1 for states that do not have phase-type sojourn distributions.
Not required for non-phase-type models.}

\item{priors}{A list specifying priors.  Each component should be
the result of a call to \code{\link{msmprior}}.  Any parameters
with priors not specified here are given default priors (normal
with mean -2 and SD 2 for log intensities, and normal with mean
0 and SD 10 for log hazard ratios).

If only one parameter is given a non-default prior, a single msmprior
call can be supplied here instead of a list.}

\item{fit_method}{Quoted name of a function from the \code{cmdstanr}
package specifying the algorithm to fit the model.  The default
\code{"sample"} uses MCMC, via \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.
Alternatives are \code{\link[cmdstanr:model-method-optimize]{cmdstanr::optimize()}},
\code{\link[cmdstanr:model-method-pathfinder]{cmdstanr::pathfinder()}}, \code{\link[cmdstanr:model-method-laplace]{cmdstanr::laplace()}} or
\code{\link[cmdstanr:model-method-variational]{cmdstanr::variational()}}.}

\item{keep_data}{Store a copy of the cleaned data in the returned
object.  \code{FALSE} by default.}

\item{...}{Other arguments to be passed to the function from
\code{cmdstanr} that fits the model.}
}
\value{
A data frame in the \code{draws} format of the
\pkg{posterior} package, containing draws from the posterior of
the model parameters.

Attributes are added to give information about the model structure,
and a class \code{"msmbayes"} is appended.

See, e.g. \code{\link{summary.msmbayes}}, \code{\link{qdf}},
\code{\link{hr}}, and similar functions, to extract parameter
estimates from the fitted model.
}
\description{
Fit a multi-state model to longitudinal data consisting of
intermittent observation of a discrete state.  Bayesian estimation
is used, via the Stan software.
}
