% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphase.r
\name{nphase}
\alias{nphase}
\alias{dnphase}
\alias{pnphase}
\alias{hnphase}
\alias{rnphase}
\title{Density, probability distribution, hazard and random
number generation functions for the Coxian
phase-type distribution with any number of phases.}
\usage{
dnphase(x, prate, arate, method = "analytic")

pnphase(x, prate, arate, method = "analytic")

hnphase(x, prate, arate, method = "analytic")

rnphase(n, prate, arate)
}
\arguments{
\item{x}{Value at which to evaluate the PDF, CDF, or hazard.}

\item{prate}{Progression rates.  Either a vector of length
\code{nphase-1}, or a matrix with \code{npar} rows and \code{nphase-1} columns.}

\item{arate}{Absorption rates.  Either a vector of length \code{nphase},
or a matrix with \code{npar} rows and \code{nphase} columns.}

\item{method}{If \code{"analytic"} then for \code{nphase} 5 or less, an
analytic solution to the matrix exponential is employed in the
calculation.  For \code{nphase} 6 or more, or if \code{method="mexp"} the
matrix exponential is determined using numerical methods, via
\code{expm::expm()}.}

\item{n}{Number of random samples to generate.}
}
\value{
A vector of length \code{n} or \code{length(x)}.
}
\description{
Density, probability distribution, hazard and random
number generation functions for the Coxian
phase-type distribution with any number of phases.
}
\details{
The number of phases, \code{nphase}, is taken from the
dimensions of the object supplied as \code{arate}.  If \code{arate} is a
vector, then the number of phases is assumed to equal the length
of this vector.  If \code{arate} is a matrix, then the number of
phases is assumed to be the number of columns.

These functions work in a vectorised way, so that alternative
parameter values or evaluation values \code{x} can be supplied.  The
number of alternative values is determined from the number of rows
\code{nrep} of \code{arate}.  Then if necessary, \code{prate} and \code{x} are
replicated to match the size of \code{arate}.
}
