% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseapprox.R
\name{phase5approx}
\alias{phase5approx}
\title{Training data to construct phase-type approximations of standard survival distributions}
\source{
Code in \code{data-raw/kl_pointwise.R} in the source package.
}
\usage{
phase5approx(family = "weibull")
}
\arguments{
\item{family}{parametric family approximated by the phase-type distribution: \code{"weibull"} or \code{"gamma"}}
}
\value{
A list with components:

\code{traindat}: data frame of training data, giving the optimal
phase-type parameters for each of a grid of shape parameters.

\code{traindat_grad} assumed gradient values at the training points,
for Hermite spline interpolation
}
\description{
Only 5-phase approximations are included.
}
