% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{phaseapprox_pars}
\alias{phaseapprox_pars}
\title{Summarise posteriors for shape and scale parameters for the sojourn distribution in a semi-Markov msmbayes model}
\usage{
phaseapprox_pars(draws, log = FALSE)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{log}{Return parameters on log scale}
}
\description{
Summarise posteriors for shape and scale parameters for the sojourn distribution in a semi-Markov msmbayes model
}
