% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{pmatrix}
\alias{pmatrix}
\title{Transition probability matrix from an msmbayes model}
\usage{
pmatrix(
  draws,
  t = 1,
  new_data = NULL,
  states = "obs",
  X = NULL,
  drop = TRUE,
  type = "posterior"
)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{t}{prediction time or vector of prediction times}

\item{new_data}{Data frame with covariate values to predict for}

\item{states}{If \code{states="obs"} (or \code{"observed"}) then this describes mean sojourn times in
the observable states.  For phase-type models this is not
generally equal to the sum of the phase-specific mean sojourn
times, because an individual may transition out of the state
before progressing to the next phase.

If \code{states="phase"} (or \code{"true"}, or \code{"latent"}) then for phase-type models, this describes mean sojourn times
in the latent state space.}

\item{X}{Lower-level alternative to specifying \code{new_data}, for
developer use only.  \code{X} is a numeric matrix formed from
column-binding the covariate design matrices for each transition
in turn.}

\item{drop}{Only used if there are no covariates supplied in
\code{new_data}. Then if \code{drop=TRUE} this returns a
\code{nstates} x \code{nstates} matrix, or if \code{drop=FALSE}
this returns a 3D array with first dimension \code{ncovs=1}.}

\item{type}{\code{"posterior"} to return \code{rvar} objects containing posterior samples.

\code{"mode"} to return posterior modes (only applicable if model was fitted
with posterior mode optimisation).}
}
\value{
Array or matrix of \code{rvar} objects giving the transition probability matrix at each requested prediction time and covariate value.  See \code{\link{qdf}} for notes on the \code{rvar} format.

For phase-type models, if \code{states="obs"}, so that we want
transition probabilities on the observable space, this returns the
probability of transition to any phase of each "destination" state,
for an individual who is in the first phase of each "starting"
state.
}
\description{
Transition probability matrix from an msmbayes model
}
\seealso{
\code{\link{pmatrixdf}} returns the same information in a tidy
data frame format.
}
