% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{pmatrixdf}
\alias{pmatrixdf}
\title{Transition probabilities from an msmbayes model, presented
as a tidy data frame}
\usage{
pmatrixdf(draws, t = 1, new_data = NULL, states = "obs")
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{t}{prediction time or vector of prediction times}

\item{new_data}{Data frame with covariate values to predict for}

\item{states}{If \code{states="obs"} (or \code{"observed"}) then this describes mean sojourn times in
the observable states.  For phase-type models this is not
generally equal to the sum of the phase-specific mean sojourn
times, because an individual may transition out of the state
before progressing to the next phase.

If \code{states="phase"} (or \code{"true"}, or \code{"latent"}) then for phase-type models, this describes mean sojourn times
in the latent state space.}
}
\value{
A data frame containing samples from the posterior distribution.
See \code{\link{qdf}} for notes on this format and how to summarise.

For phase-type models, if \code{states="obs"}, so that we want
transition probabilities on the observable space, this returns the
probability of transition to any phase of each "destination" state,
for an individual who is in the first phase of each "starting"
state.
}
\description{
Transition probabilities from an msmbayes model, presented
as a tidy data frame
}
