% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_2state_smm}
\alias{sim_2state_smm}
\title{Simulate intermittently-observed data from a semi-Markov
multi-state model with two states and reversible transitions.}
\usage{
sim_2state_smm(
  nindiv,
  obstimes,
  rfn1 = rexp,
  pars1 = list(rate = 1),
  rfn2,
  pars2
)
}
\arguments{
\item{nindiv}{number of individuals}

\item{obstimes}{observation times, common between individuals}

\item{rfn1}{Function to simulate from the sojourn distribution in
state 2.  By default this is the exponential.}

\item{pars1}{List of arguments to be passed to \code{rfn1}, specifying
parameter values for the sojourn distribution}

\item{rfn2}{Function to simulate from the sojourn distribution in
state 2.}

\item{pars2}{List of arguments to be passed to \code{rfn}, specifying
parameter values for the sojourn distribution}
}
\description{
Either or both states can have a general sojourn distribution.
}
