% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_2state_smm}
\alias{sim_2state_smm}
\title{Simulate intermittently-observed data from a semi-Markov
multi-state model with two states and reversible transitions.}
\usage{
sim_2state_smm(nindiv, obstimes, rate, rfn, pars)
}
\arguments{
\item{nindiv}{number of individuals}

\item{obstimes}{observation times, common between individuals}

\item{rate}{Rate of the exponential sojourn distribution for state 1.}

\item{rfn}{Function to simulate from the sojourn distribution in
state 2.}

\item{pars}{List of arguments to be passed to \code{rfn}, specifying
parameter values for the sojourn distribution}
}
\description{
State 1 has an exponential sojourn distribution, and state 2 has a
more general sojourn distribution.
}
