% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{soj_prob}
\alias{soj_prob}
\title{Sojourn probability in a state of a msmbayes model}
\usage{
soj_prob(draws, t, state, new_data = NULL, method = "analytic")
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{t}{Time since state entry}

\item{state}{State of interest (integer)}

\item{new_data}{Data frame with covariate values to predict for}

\item{method}{If \code{"analytic"} then for \code{nphase} 5 or less, an
analytic solution to the matrix exponential is employed in the
calculation.  For \code{nphase} 6 or more, or if \code{method="mexp"} the
matrix exponential is determined using numerical methods, via
\code{expm::expm()}.}
}
\value{
A data frame with column \code{value} giving the probability of
remaining in \code{state} by time \code{t} since state entry, as an \code{rvar}
object. Other columns give the time and any covariate values.

See \code{\link{qdf}} for notes on the \code{rvar} format.
}
\description{
Sojourn probability in a state of a msmbayes model
}
