% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{soj_prob}
\alias{soj_prob}
\title{Sojourn probability in a state of a msmbayes model}
\usage{
soj_prob(draws, t, state, new_data = NULL)
}
\arguments{
\item{draws}{Object returned by \code{\link{msmbayes}}.}

\item{t}{Time since state entry}

\item{state}{State of interest (integer)}

\item{new_data}{Data frame with covariate values to predict for}
}
\value{
A data frame with column \code{value} giving the probability of
remaining in \code{state} by time \code{t} since state entry, as an \code{rvar}
object. Other columns give the time and any covariate values.

See \code{\link{qdf}} for notes on the \code{rvar} format.
}
\description{
Sojourn probability in a state of a msmbayes model
}
