% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statetable.R
\name{statetable}
\alias{statetable}
\title{Summarise intermittenly-observed multi-state data}
\usage{
statetable(
  data,
  state = "state",
  subject = "subject",
  time = "time",
  time_groups = 1
)
}
\arguments{
\item{data}{Data frame giving the observed data.}

\item{state}{Character string naming the observed state variable in
the data.  This variable must either be an integer in 1,2,...,K,
where K is the number of states, or a factor with these integers
as level labels.  If omitted, this is assumed to be \code{"state"}.}

\item{subject}{Character string naming the individual ID variable in the data.
If omitted, this is assumed to be \code{"subject"}.}

\item{time}{Character string naming the observation time variable in the data.
If omitted, this is assumed to be \code{"time"}.}

\item{time_groups}{Number of groups to summarise the time
intervals by.  The transitions are categorised into groups
according to equally-spaced quantiles of the time interval
length.}
}
\value{
A data frame with columns \code{fromstate}, \code{tostate}, \code{timelag} and
\code{n} (count of transitions)
}
\description{
Tabulate observed transitions between states over successive observations, by
from-state, to-state and (optionally) time interval length.
}
\details{
This is like the function statetable.msm in msm, except that it
uses msmbayes syntax for specifying the data, it also summarises
the length of the time intervals between successive observations,
and returns a tidy data frame.
}
