% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmbayes.R
\name{summary.msmbayes}
\alias{summary.msmbayes}
\title{Summarise basic parameter estimates from an msmbayes model}
\usage{
\method{summary}{msmbayes}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{msmbayes}}.}

\item{...}{Further arguments passed to both \code{\link{qdf}} and \code{\link{loghr}}.}
}
\value{
A data frame with one row for each parameter.  The
posterior distribution for the parameter is encoded in the column
\code{value}, which has the \code{rvar} data type defined by the
\pkg{posterior package}.  This can be summarised by calling
\code{summary} again on the data frame (see the examples).

\code{q} are the transition intensities (at covariate values of zero)

Remaining parameters (in non-HMMs) are log hazard ratios for
covariate effects.
}
\description{
Summarise basic parameter estimates from an msmbayes model
}
\examples{
summary(infsim_model)
summary(summary(infsim_model))
summary(summary(infsim_model), median, ~quantile(.x, 0.025, 0.975))

}
\seealso{
\code{\link{qdf}}, \code{\link{hr}}, \code{\link{loghr}},
\code{\link[posterior:summarise_draws]{posterior::summarise_draws}}
}
